Added pid controller for velocity

+++ Inc/PID_base.h
@@ -11,64 +11,64 @@
 
 typedef struct PID{
 	//Controller parameters
-	double P;
-	double I;
-	double D;
-	double F;
+	float P;
+	float I;
+	float D;
+	float F;
 
 	//Limits
-	double maxIOutput;
-	double minIOutput;
-	double maxError;
-	double errorSum;
-	double prevError;
-	double deadTime;
-	double frequency;
+	float maxIOutput;
+	float minIOutput;
+	float maxError;
+	float errorSum;
+	float prevError;
+	float deadTime;
+	float frequency;
 
-	double maxOutput;
-	double minOutput;
+	float maxOutput;
+	float minOutput;
 
-	double setpoint;
+	float setpoint;
 
-	double lastActual;
+	float lastActual;
 
 	//Flags
 	int firstRun;
 	int reversed;
 
 	//Ramping and descent limits
-	double outputRampRate;
-	double outputDescentRate;
-	double lastOutput;
+	float outputRampRate;
+	float outputDescentRate;
+	float lastOutput;
 
-	double outputFilter;
+	float outputFilter;
 
-	double setpointRange;
+	float setpointRange;
 
 	uint32_t prev_time;
 } PID_Struct;
 
 void PID_Init(PID_Struct* pid);
 void checkSigns(PID_Struct* pid);
-double clamp(double value, double min, double max);
-uint8_t bounded(double value, double min, double max);
-void PID_setPID(PID_Struct* pid, double p, double i, double d);
-void PID_setPIDF(PID_Struct* pid, double p, double i, double d, double f);
-void PID_setF(PID_Struct* pid, double f);
-void PID_setMaxIOutput(PID_Struct* pid, double maximum);
-void PID_setMinIOutput(PID_Struct* pid, double minimum);
-void PID_setOutputLimits(PID_Struct* pid, double min, double max);
+float clamp(float value, float min, float max);
+uint8_t bounded(float value, float min, float max);
+void PID_setPID(PID_Struct* pid, float p, float i, float d);
+void PID_setPIDF(PID_Struct* pid, float p, float i, float d, float f);
+void PID_setF(PID_Struct* pid, float f);
+void PID_setMaxIOutput(PID_Struct* pid, float maximum);
+void PID_setMinIOutput(PID_Struct* pid, float minimum);
+void PID_setOutputLimits(PID_Struct* pid, float min, float max);
 void PID_setDirection(PID_Struct* pid, int reversed);
-void PID_setSetpoint(PID_Struct* pid, double setpoint);
-double PID_skipCycle(PID_Struct* pid);
-double PID_getOutput(PID_Struct* pid, double actual, double setpoint);
-double PID_getOutputFast(PID_Struct* pid);
+void PID_setSetpoint(PID_Struct* pid, float setpoint);
+float PID_skipCycle(PID_Struct* pid);
+float PID_getOutput(PID_Struct* pid, float actual, float setpoint);
+float PID_getOutputFast(PID_Struct* pid);
 void PID_reset(PID_Struct* pid);
-void PID_setOutputRampRate(PID_Struct* pid, double rate);
-void PID_setOutputDescentRate(PID_Struct* pid, double rate);
-void PID_setSetpointRange(PID_Struct* pid, double range);
-void PID_setOutputFilter(PID_Struct* pid, double strength);
-void PID_setFrequency(PID_Struct* pid, double freq);
+void PID_setOutputRampRate(PID_Struct* pid, float rate);
+void PID_setOutputDescentRate(PID_Struct* pid, float rate);
+void PID_setSetpointRange(PID_Struct* pid, float range);
+void PID_setOutputFilter(PID_Struct* pid, float strength);
+void PID_setFrequency(PID_Struct* pid, float freq);
 
 #endif /* INC_PID_H_ */
